refactor(model): give the model-provider contracts a home below llms
`runtime` names `llms` in 34 files, and `llms` is what puts `candle`, `mistralrs`,
`tokenizers`, `gemm` and every vendor SDK on `runtime`'s prerequisite path — 163
crates that must finish before `runtime` starts compiling. Splitting the providers
out is what removes them, and this is the contract they need first.

`Chat`, `Embed` and `Rerank` move into three crates rather than one, because their
dependency sets are nearly disjoint:

  rerank-api   async-trait, snafu
  embed-api    + async-openai, cache, chunking, hf-hub, runtime-rate-control
  chat-api     + async-openai, async-stream, futures, spicepod, serde_json

A reranker vendor is a small HTTP client and has no business waiting on `cache` or
`hf-hub`, which only the embedding contract needs — and a change to `Chat` should not
rebuild rerank providers. The capability boundary is already real at runtime, where
`RerankerModelStore` is a separate map from the chat store; this makes it real at
build time. Splitting per capability *and* per vendor was considered and rejected:
Bedrock implements chat and embed on the same AWS SDK, so that split would shed
nothing and give the client construction two homes.

`chat-api` and `rerank-api` sit at foundation; their whole closure is external crates
plus `spicepod`, and keeping them there is the constraint — a contract that grows a
shared-utility dependency has stopped being a thin interface. `embed-api` cannot,
because `Embed`'s default bodies name `cache` and `chunking`.

`llms` depends on all three and re-exports each with a glob, so no caller changes.
That is scaffolding, not the end state: while `llms::chat::Chat` resolves, nothing
forces `runtime` to name `chat-api`, and a public glob launders the contract past the
layering guard. Demoting those globs is the next step, and it is what makes the
compiler enumerate the call sites that still route through `llms`.

Measured after this change: `runtime`'s closure is 1,519 crates with 166 still
reachable only via `llms`. The shed is deliberately zero here — nothing moves until
`model-local-chat` and `model-local-embed` peel off the mistralrs and candle stacks,
which this contract is the prerequisite for.
