+++ Cargo.toml
@@ -30,6 +30,9 @@ members = [
     "crates/google-genai",
     "crates/hash-index",
     "crates/llms",
+    "crates/chat-api",
+    "crates/embed-api",
+    "crates/rerank-api",
     "crates/ns_lookup",
     "crates/object_store_occ",
     "crates/otel-arrow",
