+++ crates/chat-api/Cargo.toml
@@ -1,0 +1,42 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "chat-api"
+description = "The chat-completion contract: the `Chat` trait a model provider implements, the NSQL surface, the vocabulary they are expressed in, and the errors they report."
+edition.workspace = true
+exclude.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+async-openai.workspace = true
+async-stream.workspace = true
+async-trait.workspace = true
+futures.workspace = true
+rand.workspace = true
+serde_json.workspace = true
+snafu.workspace = true
+spicepod = { path = "../spicepod" }
+tokio.workspace = true
+tracing.workspace = true
+tracing-futures.workspace = true
+
+[lints]
+workspace = true
