+++ crates/embed-api/Cargo.toml
@@ -1,0 +1,40 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "embed-api"
+description = "The embedding contract: the `Embed` trait an embedding provider implements, the vocabulary it is expressed in, and the errors it reports."
+edition.workspace = true
+exclude.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+async-openai.workspace = true
+async-trait.workspace = true
+cache = { path = "../cache" }
+chunking = { path = "../chunking" }
+hf-hub = { workspace = true }
+runtime-rate-control = { path = "../runtime-rate-control" }
+snafu.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
