+++ crates/llms/Cargo.toml
@@ -13,6 +13,9 @@ version.workspace = true
 snafu.workspace = true
 
 async-openai.workspace = true
+chat-api = { path = "../chat-api" }
+embed-api = { path = "../embed-api" }
+rerank-api = { path = "../rerank-api" }
 async-stream.workspace = true
 async-trait.workspace = true
 aws-config.workspace = true
